(QENG-1640) Limit number of facts generate
Previous to this commit, the random facts generated per agent would
increase with each puppet agent run, due to the fact names also being
randomly created.

This commit changes it to generate consistent fact names, but allow the
user to configure the number of them to see where the limits are.

Further, it will also randomize the data for the facts on each run and
vary the length of the facts, to simulate users having facts like SSH
keys or other random data.